Fix missing field

+++ F1_DB/dbo/StoreProcedures/sp_InsertDriver.sql
@@ -1,7 +1,7 @@
 ﻿CREATE PROCEDURE [dbo].[sp_InsertDriver]
 	@FirstName nvarchar(50),
 	@LastName nvarchar(50),
-	@Number smallint,
+	@Number tinyint,
     @PhotoPath nvarchar(260)
 AS
 	INSERT INTO Driver
